revision: testing, improved completion and various small bugfixes

+++ apt.zsh
@@ -5,24 +5,30 @@ PACKAGEMANAGER_FZF_APT_PREVIEW='apt-cache show {}'
 ###########################
 
 function _fzf_complete_apt() {
-    local arguments=$@
+    local arguments="$@"
 
     if [[ $arguments == 'apt install'* ]]; then
-        _fzf_complete -m --preview $PACKAGEMANAGER_FZF_APT_PREVIEW -- "$@" < <(apt-cache search . | cut -d' ' -f1)
+        _fzf_complete -m --header="apt install" --preview "$PACKAGEMANAGER_FZF_APT_PREVIEW" -- "$@" < <(apt-cache search . | cut -d' ' -f1 | grep .)
+    elif [[ $arguments == 'apt reinstall'* ]]; then
+        _fzf_complete -m --header="apt reinstall" --preview "$PACKAGEMANAGER_FZF_APT_PREVIEW" -- "$@" < <(dpkg --get-selections | grep -w install | cut -f1 | grep .)
     elif [[ $arguments == 'apt remove'* ]]; then
-        _fzf_complete -m --preview $PACKAGEMANAGER_FZF_APT_PREVIEW -- "$@" < <(dpkg --get-selections | cut -f1)
+        _fzf_complete -m --header="apt remove" --preview "$PACKAGEMANAGER_FZF_APT_PREVIEW" -- "$@" < <(dpkg --get-selections | grep -w install | cut -f1 | grep .)
+    elif [[ $arguments == 'apt purge'* ]]; then
+        _fzf_complete -m --header="apt purge" --preview "$PACKAGEMANAGER_FZF_APT_PREVIEW" -- "$@" < <(dpkg --get-selections | grep -w install | cut -f1 | grep .)
     else
         eval "zle ${fzf_default_completion:-expand-or-complete}"
     fi
 }
 
 function _fzf_complete_apt-get() {
-    local arguments=$@
+    local arguments="$@"
 
     if [[ $arguments == 'apt-get install'* ]]; then
-        _fzf_complete -m --preview $PACKAGEMANAGER_FZF_APT_PREVIEW -- "$@" < <(apt-cache search . | cut -d' ' -f1)
+        _fzf_complete -m --header="apt-get install" --preview "$PACKAGEMANAGER_FZF_APT_PREVIEW" -- "$@" < <(apt-cache search . | cut -d' ' -f1 | grep .)
     elif [[ $arguments == 'apt-get remove'* ]]; then
-        _fzf_complete -m --preview $PACKAGEMANAGER_FZF_APT_PREVIEW -- "$@" < <(dpkg --get-selections | cut -f1)
+        _fzf_complete -m --header="apt-get remove" --preview "$PACKAGEMANAGER_FZF_APT_PREVIEW" -- "$@" < <(dpkg --get-selections | grep -w install | cut -f1 | grep .)
+    elif [[ $arguments == 'apt-get purge'* ]]; then
+        _fzf_complete -m --header="apt-get purge" --preview "$PACKAGEMANAGER_FZF_APT_PREVIEW" -- "$@" < <(dpkg --get-selections | grep -w install | cut -f1 | grep .)
     else
         eval "zle ${fzf_default_completion:-expand-or-complete}"
     fi
@@ -32,23 +38,23 @@ function _fzf_complete_apt-get() {
 # standalone functions #
 ########################
 
-function __pmf_apt_install() {
-    local packages=("${(@f)$(apt-cache search . | cut -d' ' -f1 | fzf --query="$1" -m --preview $PACKAGEMANAGER_FZF_APT_PREVIEW)}")
+function __fpm_apt_install() {
+    local packages=("${(@f)$(apt-cache search . | cut -d' ' -f1 | grep . | fzf --query="$1" -m --header="apt install" --preview "$PACKAGEMANAGER_FZF_APT_PREVIEW")}")
 
-    if [[ "${packages[@]}" ]]; then
+    if (( ${#packages} )) && [[ -n "${packages[1]}" ]]; then
         sudo apt-get install "${packages[@]}"
     fi
 }
 
-function __pmf_apt_remove() {
-    local packages=("${(@f)$(dpkg --get-selections | cut -f1 | fzf --query="$1" -m --preview $PACKAGEMANAGER_FZF_APT_PREVIEW)}")
+function __fpm_apt_remove() {
+    local packages=("${(@f)$(dpkg --get-selections | grep -w install | cut -f1 | grep . | fzf --query="$1" -m --header="apt remove" --preview "$PACKAGEMANAGER_FZF_APT_PREVIEW")}")
 
-    if [[ "${packages[@]}" ]]; then
+    if (( ${#packages} )) && [[ -n "${packages[1]}" ]]; then
         sudo apt-get remove "${packages[@]}"
     fi
 }
 
-function __pmf_apt_setup() {
-    alias aip=__pmf_apt_install
-    alias arp=__pmf_apt_remove
+function __fpm_apt_setup() {
+    alias aip=__fpm_apt_install
+    alias arp=__fpm_apt_remove
 }
